fix(usb): 修复字符串描述符写入时的类型转换问题
将字符串大小乘以 char16_t 字节长度的结果显式转换为 int64_t 类型，
避免潜在的整数溢出问题，并添加缺失的 cstdint 头文件引用。

+++ include/base/embedded/usb/descriptor/StringDescriptorWriter.h
@@ -4,6 +4,7 @@
 #include "base/string/define.h"
 #include "DescriptorWriter.h"
 #include <cstddef>
+#include <cstdint>
 #include <stdexcept>
 #include <string>
 
@@ -62,7 +63,7 @@ namespace base
 			{
 				base::ReadOnlySpan span{
 					reinterpret_cast<uint8_t const *>(str.data()),
-					str.size() * sizeof(char16_t),
+					static_cast<int64_t>(str.size() * sizeof(char16_t)),
 				};
 
 				_writer.WriteDataLength(span.Size());
